db jpa에 의존하여 재생성 및 대학교 서치 체크

+++ src/main/java/muckkitlist_spring/muckkitlist_spring/entity/RestaurantInfoEntity.java
@@ -22,7 +22,7 @@ public class RestaurantInfoEntity {
   private String address;
 
   @Column(name = "review_count")
-  private long reviewCount;
+  private int reviewCount;
 
   @Column(name = "avg_grade")
   private double avgGrade;
